Update clip id's to longs. Add penalty, shot and stoppage types

+++ Entities/DbModels/GamePlayEvents/DbGoal.cs
@@ -26,8 +26,8 @@ public class DbGoal : IDbGameEvent
     public int ScoringPlayerId { get; set; }
     public int? GoalieId { get; set; }
     public string HighlightClipSharingUrl { get; set; } = string.Empty;
-    public int HighlightClipId { get; set; }
-    public double DiscreetClipId { get; set; }
+    public long HighlightClipId { get; set; }
+    public long DiscreetClipId { get; set; }
     public string PptReplayUrl { get; set; } = string.Empty;
     [ForeignKey(nameof(GameId))]
     public DbGameRaw? Game { get; set; }
